UIMA-4474 - fixed getter call for numbers - change order
git-svn-id: https://svn.apache.org/repos/asf/uima/ruta/trunk@1687007 13f79535-47bb-0310-9956-ffa450edef68

+++ ruta-core/src/main/java/org/apache/uima/ruta/expression/feature/GenericFeatureExpression.java
@@ -111,14 +111,14 @@ public class GenericFeatureExpression extends RutaExpression implements INumberE
     if(numberExpression == null) {
       numberExpression = ExpressionFactory.createNumberFeatureExpression(featureExpression);
     }
-    return numberExpression.getIntegerValue(parent, annotation, stream);
+    return numberExpression.getDoubleValue(parent, annotation, stream);
   }
 
   public float getFloatValue(RutaBlock parent, AnnotationFS annotation, RutaStream stream) {
     if(numberExpression == null) {
       numberExpression = ExpressionFactory.createNumberFeatureExpression(featureExpression);
     }
-    return numberExpression.getIntegerValue(parent, annotation, stream);
+    return numberExpression.getFloatValue(parent, annotation, stream);
   }
